git-svn-id: svn://localhost/bt1/trunk@23 f571bf5b-3fff-47ac-b1e0-9c329ff6e91c

+++ BrawlLib/System/Windows/Forms/ModelPanel.cs
@@ -142,7 +142,7 @@ namespace System.Windows.Forms
 
         protected override void OnMouseWheel(MouseEventArgs e)
         {
-            int z = e.Delta / 120;
+            float z = (float)e.Delta / 120;
             if (Control.ModifierKeys == Keys.Shift)
                 z *= 32;
